# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_leleo886_lopic_internal_error.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.BackupTask:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      error:
        type: string
      id:
        type: integer
      size:
        type: integer
      start_time:
        type: string
      status:
        type: string
      storage_path:
        type: string
      updated_at:
        type: string
    type: object
...
  models.RestoreTask:
    properties:
      backup_task:
        $ref: '#/definitions/models.BackupTask'
      backup_task_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      error:
        type: string
      id:
        type: integer
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  success.DataResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  success.SuccessResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API
  title: Lopic API
  version: 1.0.0
paths:
...
  /api/admin/backup:
    post:
      consumes:
      - application/json
      description: 创建系统全量备份
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.BackupTask'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 创建备份
      tags:
      - backup
  /api/admin/backup/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的备份
      parameters:
      - description: 备份ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除备份
      tags:
      - backup
  /api/admin/backup/download/{id}:
    get:
      consumes:
      - application/json
      description: 下载指定的备份文件
      parameters:
      - description: 备份ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 备份文件
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 下载备份文件
      tags:
      - backup
  /api/admin/backup/list:
    get:
      consumes:
      - application/json
      description: 获取所有备份任务列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.BackupTask'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取备份列表
      tags:
      - backup
  /api/admin/backup/restore/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的恢复任务记录
      parameters:
      - description: 恢复任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/success.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除恢复任务
      tags:
      - backup
    post:
      consumes:
      - application/json
      description: 从指定备份恢复系统
      parameters:
      - description: 备份ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.RestoreTask'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 恢复备份
      tags:
      - backup
  /api/admin/backup/restore/list:
    get:
      consumes:
      - application/json
      description: 获取所有恢复任务记录
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RestoreTask'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取恢复记录
      tags:
      - backup
  /api/admin/backup/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传备份文件压缩包，上传完成后会添加记录到备份表中，上传逻辑异步
      parameters:
      - description: 备份文件压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/success.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.BackupTask'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_leleo886_lopic_internal_error.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: 上传备份文件
      tags:
      - backup
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"